+++ architecture_tool_django/dashboard/templatetags/extra_filters.py
@@ -4,3 +4,8 @@ from django.template.defaulttags import register
 @register.filter
 def get_item(dictionary, key):
     return dictionary.get(key)
+
+
+@register.filter
+def understored_name(value):
+    return value.replace("-", "_")